move data loading util
