+++ behavenet/data/utils.py
@@ -4,6 +4,8 @@ import os
 import numpy as np
 import pickle
 
+from behavenet.fitting.utils import export_session_info_to_csv
+
 
 def get_data_generator_inputs(hparams, sess_ids, check_splits=True):
     """Helper function for generating signals, transforms and paths.
@@ -280,6 +282,52 @@ def get_data_generator_inputs(hparams, sess_ids, check_splits=True):
     return hparams, signals_list, transforms_list, paths_list
 
 
+def build_data_generator(hparams, sess_ids, export_csv=True):
+    """Helper function to build data generator from hparams dict.
+
+    Parameters
+    ----------
+    hparams : :obj:`dict`
+        needs to contain information specifying data inputs to model
+    sess_ids : :obj:`list` of :obj:`dict`
+        each entry is a session dict with keys 'lab', 'expt', 'animal', 'session'
+    export_csv : :obj:`bool`, optional
+        export csv file containing session info (useful when fitting multi-sessions)
+
+    Returns
+    -------
+    :obj:`ConcatSessionsGenerator` object
+        data generator
+
+    """
+    from behavenet.data.data_generator import ConcatSessionsGenerator
+    print('using data from following sessions:')
+    for ids in sess_ids:
+        print('%s' % os.path.join(
+            hparams['save_dir'], ids['lab'], ids['expt'], ids['animal'], ids['session']))
+    hparams, signals, transforms, paths = get_data_generator_inputs(hparams, sess_ids)
+    if hparams.get('trial_splits', None) is not None:
+        # assumes string of form 'train;val;test;gap'
+        trs = [int(tr) for tr in hparams['trial_splits'].split(';')]
+        trial_splits = {'train_tr': trs[0], 'val_tr': trs[1], 'test_tr': trs[2], 'gap_tr': trs[3]}
+    else:
+        trial_splits = None
+    print('constructing data generator...', end='')
+    data_generator = ConcatSessionsGenerator(
+        hparams['data_dir'], sess_ids,
+        signals_list=signals, transforms_list=transforms, paths_list=paths,
+        device=hparams['device'], as_numpy=hparams['as_numpy'], batch_load=hparams['batch_load'],
+        rng_seed=hparams['rng_seed_data'], trial_splits=trial_splits,
+        train_frac=hparams['train_frac'])
+    # csv order will reflect dataset order in data generator
+    if export_csv:
+        export_session_info_to_csv(os.path.join(
+            hparams['expt_dir'], str('version_%i' % hparams['version'])), sess_ids)
+    print('done')
+    print(data_generator)
+    return data_generator
+
+
 def check_same_training_split(model_path, hparams):
     """Ensure data rng seed and trial splits are same for two models."""
 
@@ -501,7 +549,6 @@ def load_labels_like_latents(hparams, sess_ids, sess_idx, data_key='labels'):
 
     """
     import copy
-    from behavenet.fitting.utils import build_data_generator
 
     hparams_new = copy.deepcopy(hparams)
     hparams_new['model_class'] = data_key
